// osa9/flight-diary/frontend/src/utils/types.tsx
export enum Weather {
    Sunny = 'sunny',
    Rainy = 'rainy',
    Cloudy = 'cloudy',
    Stormy = 'stormy',
    Windy = 'windy',
}
  
export enum Visibility {
    Great = 'great',
    Good = 'good',
    Ok = 'ok',
    Poor = 'poor',
}
export type Message={
    message:string|null
}
export interface Diary {
    id: number;
    date: string;
    weather: Weather;
    visibility: Visibility;
    comment: string;
}
  
export type NewDiary = Omit<Diary, 'id'>;
  
export type NonSensitiveDiary = Omit<Diary, 'comment'>;